requirements_pi.txt
# Raspberry Pi 5 optimized requirements
# Install with: pip install -r requirements_pi.txt

# Core dependencies
google-genai>=0.3.0
python-dotenv>=1.0.0
requests>=2.31.0
httpx>=0.25.0  # Async HTTP client with better timeout handling

# Audio
pyaudio>=0.2.13
sounddevice>=0.4.6
numpy>=1.24.0,<2.0.0

# Wake word detection
pvporcupine>=3.0.0

# TTS (local, fast)
piper-tts>=1.2.0

# VAD - use ONNX version for lower memory on Pi
# torch is heavy, consider onnxruntime instead for production
onnxruntime>=1.16.0

# Smart home
pyControl4>=1.1.0

# Music (optional)
ytmusicapi>=1.3.0
yt-dlp>=2023.12.0

# UI Gateway (optional)
fastapi>=0.104.0
uvicorn>=0.24.0
websockets>=12.0

# Utilities
pydantic>=2.5.0
tenacity>=8.2.0  # Retry logic

# Development/debugging (optional, comment out for production)
# py-spy>=0.3.14
# memory-profiler>=0.61.0
